Remove epmc from OTConfig and add to LiteraureSection in Configuration.scala Add literature step to evidence's dependencies list in reference.conf Move epmc to literature in reference.conf Remove epmc running command from Main.scala
