+++ src/main/resources/reference.conf
@@ -42,7 +42,7 @@ etl-dag {
     {step: "target", dependencies: ["reactome"]},
     {step: "interaction", dependencies: ["target"]},
     {step: "targetValidation", dependencies: ["target"]},
-    {step: "evidence", dependencies: ["disease", "target"]},
+    {step: "evidence", dependencies: ["disease", "target", "literature"]},
     {step: "association", dependencies: ["evidence", "disease"]},
     {step: "associationOTF", dependencies: ["evidence", "target", "disease", "reactome"]},
     {step: "search", dependencies: ["target", "drug", "evidence", "association", "disease"]},
@@ -76,59 +76,6 @@ common {
 
 }
 
-epmc {
-  input {
-    cooccurences = ${literature.processing.outputs.cooccurrences}
-  }
-  output {
-    format = "json" // to be consistent with the other Evidence files
-    path = ${common.input}"/evidence-files/epmc"
-    options = [
-      {k: "compression", v: "gzip"},
-    ]
-  }
-  excluded-target-terms = [
-    "TEC",
-    "TECS",
-    "Tec",
-    "tec",
-    "'",
-    "(",
-    ")",
-    "-",
-    "-S",
-    "S",
-    "S-",
-    "SS",
-    "SSS",
-    "Ss",
-    "Ss-",
-    "s",
-    "s-",
-    "ss",
-    "U3",
-    "U6",
-    "u6",
-    "SNORA70",
-    "U2",
-    "U8"
-  ]
-  # The following are the relevant sections for disease/target associations as described in PMID28587637
-  sections-of-interest = [
-    "title",
-    "abstract",
-    "intro",
-    "case",
-    "figure",
-    "table",
-    "discuss",
-    "concl",
-    "results",
-    "appendix",
-    "other",
-  ]
-  print-metrics = true
-}
 reactome {
   inputs {
     pathways {
@@ -1420,6 +1367,60 @@ literature {
       format = ${common.output-format}
       path = ${literature.output}"/vectors"
     }
+  }
+
+  epmc {
+    input {
+      cooccurences = ${literature.processing.outputs.cooccurrences}
+    }
+    output {
+      format = "json" // to be consistent with the other Evidence files
+      path = ${common.input}"/evidence-files/epmc"
+      options = [
+        {k: "compression", v: "gzip"},
+      ]
+    }
+    excluded-target-terms = [
+      "TEC",
+      "TECS",
+      "Tec",
+      "tec",
+      "'",
+      "(",
+      ")",
+      "-",
+      "-S",
+      "S",
+      "S-",
+      "SS",
+      "SSS",
+      "Ss",
+      "Ss-",
+      "s",
+      "s-",
+      "ss",
+      "U3",
+      "U6",
+      "u6",
+      "SNORA70",
+      "U2",
+      "U8"
+    ]
+    # The following are the relevant sections for disease/target associations as described in PMID28587637
+    sections-of-interest = [
+      "title",
+      "abstract",
+      "intro",
+      "case",
+      "figure",
+      "table",
+      "discuss",
+      "concl",
+      "results",
+      "appendix",
+      "other",
+    ]
+    print-metrics = true
   }
 }
