+++ src/main/scala/io/opentargets/etl/backend/literature/Epmc.scala
@@ -1,20 +1,10 @@
-package io.opentargets.etl.backend
+package io.opentargets.etl.backend.literature
 
 import com.typesafe.scalalogging.LazyLogging
-import io.opentargets.etl.backend.spark.{IOResource, IoHelpers}
+import io.opentargets.etl.backend.ETLSessionContext
 import io.opentargets.etl.backend.spark.IoHelpers.IOResources
-import org.apache.spark.sql.functions.{
-  coalesce,
-  col,
-  collect_set,
-  length,
-  struct,
-  sum,
-  trim,
-  when,
-  size,
-  lit
-}
+import io.opentargets.etl.backend.spark.{IOResource, IoHelpers}
+import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types.StringType
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
@@ -25,7 +15,7 @@ object Epmc extends LazyLogging {
 
     logger.info("Prepare EPMC evidence")
 
-    val conf = context.configuration.epmc
+    val conf = context.configuration.literature.epmc
 
     val inputDataFrames = IoHelpers.readFrom(
       Map(
