+++ src/main/scala/io/opentargets/etl/backend/literature/Literature.scala
@@ -41,6 +41,9 @@ object Literature extends LazyLogging {
     Embedding()
     logger.info("Run literature vectors")
     Vectors()
+    logger.info("Run EPMC")
+    Epmc()
+
   }
 
 }
